When doing the clean up, skip entries that do not have a fileRef attribute
Entries that do not have this entry are broken, but they seem to happen in real life.

This should fix issue #21

+++ src/pbxproj/pbxobjects.py
@@ -40,6 +40,8 @@ class PBXProjFile(DictionaryBoundObject):
         files = self._objects.getobjects(isa="PBXBuildFile")
         removed_files = []
         for (identifier, file) in files:
+            if not file.has_attr("fileRef"):
+                continue
             if not file.fileRef in self._objects:
                 removed_files.append(identifier)
                 self._objects.delete(identifier)
